+++ Tests/EditModeTests/Authoring/PrefabConversionTests.cs
@@ -12,7 +12,9 @@ namespace Unity.Physics.Tests.Authoring
 {
     class PrefabConversionTests : PrefabConversionTestsBase
     {
-        [Ignore("Failing in Packageworks, https://jira.unity3d.com/browse/DOTS-10120")]
+
+        #if false
+        //Temporarily disabling this test due to failures exclusive to Packageworks
         [Test]
         public void PrefabConversion_ChildCollider_ForceUnique([Values] bool forceUniqueCollider)
         {
@@ -25,7 +27,8 @@ namespace Unity.Physics.Tests.Authoring
                     if (forceUniqueCollider) rigidBody.AddComponent<ForceUniqueColliderAuthoring>();
                 });
         }
-
+        #endif
+        
         [Test]
         public void PrefabConversion_GameObjectPrefabInstances_ForceUnique([Values] bool forceUniqueCollider)
         {